moving ad code to wrapper
