+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"/Users/wleaverton/dev/jwplayer-fandom/src/i18n.ts":"1","/Users/wleaverton/dev/jwplayer-fandom/src/main.ts":"2","/Users/wleaverton/dev/jwplayer-fandom/src/players/DesktopArticleVideoPlayer/DesktopArticleVideoPlayer.tsx":"3","/Users/wleaverton/dev/jwplayer-fandom/src/players/DesktopArticleVideoPlayer/UnmuteButton.tsx":"4","/Users/wleaverton/dev/jwplayer-fandom/src/players/DesktopArticleVideoPlayer/UserFeedback/CloseButton.tsx":"5","/Users/wleaverton/dev/jwplayer-fandom/src/players/DesktopArticleVideoPlayer/UserFeedback/ThumbDownButton.tsx":"6","/Users/wleaverton/dev/jwplayer-fandom/src/players/DesktopArticleVideoPlayer/UserFeedback/ThumbUpButton.tsx":"7","/Users/wleaverton/dev/jwplayer-fandom/src/players/DesktopArticleVideoPlayer/UserFeedback/UserFeedback.tsx":"8","/Users/wleaverton/dev/jwplayer-fandom/src/players/DesktopArticleVideoPlayer/VideoDetails.tsx":"9","/Users/wleaverton/dev/jwplayer-fandom/src/players/VideoPlayer.tsx":"10","/Users/wleaverton/dev/jwplayer-fandom/src/players/shared/JwPlayerWrapper.tsx":"11","/Users/wleaverton/dev/jwplayer-fandom/src/players/shared/PlayerContext.ts":"12","/Users/wleaverton/dev/jwplayer-fandom/src/players/shared/PlayerWrapper.tsx":"13","/Users/wleaverton/dev/jwplayer-fandom/src/plugins/fandom-wirewax.plugin.ts":"14","/Users/wleaverton/dev/jwplayer-fandom/src/types.d.ts":"15","/Users/wleaverton/dev/jwplayer-fandom/src/utils/communication/communicationService.js":"16","/Users/wleaverton/dev/jwplayer-fandom/src/utils/communication/index.js":"17","/Users/wleaverton/dev/jwplayer-fandom/src/utils/communication/ofType.js":"18","/Users/wleaverton/dev/jwplayer-fandom/src/utils/useOnScreen.ts":"19","/Users/wleaverton/dev/jwplayer-fandom/src/players/shared/AdWrapper.tsx":"20"},{"size":7312,"mtime":1645827956782,"results":"21","hashOfConfig":"22"},{"size":301,"mtime":1645837428879,"results":"23","hashOfConfig":"22"},{"size":2878,"mtime":1646250015371,"results":"24","hashOfConfig":"22"},{"size":1243,"mtime":1646094534135,"results":"25","hashOfConfig":"22"},{"size":396,"mtime":1646087596123,"results":"26","hashOfConfig":"22"},{"size":1039,"mtime":1646087596124,"results":"27","hashOfConfig":"22"},{"size":1035,"mtime":1646087596124,"results":"28","hashOfConfig":"22"},{"size":968,"mtime":1646094550385,"results":"29","hashOfConfig":"22"},{"size":897,"mtime":1646087596126,"results":"30","hashOfConfig":"22"},{"size":171,"mtime":1646087596127,"results":"31","hashOfConfig":"22"},{"size":1696,"mtime":1646087596128,"results":"32","hashOfConfig":"22"},{"size":240,"mtime":1645838027635,"results":"33","hashOfConfig":"22"},{"size":469,"mtime":1646087596129,"results":"34","hashOfConfig":"22"},{"size":6083,"mtime":1646087596130,"results":"35","hashOfConfig":"22"},{"size":1570,"mtime":1646086252614,"results":"36","hashOfConfig":"22"},{"size":665,"mtime":1646182736037,"results":"37","hashOfConfig":"22"},{"size":102,"mtime":1646183247638,"results":"38","hashOfConfig":"22"},{"size":191,"mtime":1646182588793,"results":"39","hashOfConfig":"22"},{"size":806,"mtime":1645837428891,"results":"40","hashOfConfig":"22"},{"size":1259,"mtime":1646250495914,"results":"41","hashOfConfig":"22"},{"filePath":"42","messages":"43","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"149w94r",{"filePath":"44","messages":"45","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"46","messages":"47","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"48","messages":"49","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"50","messages":"51","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"52","messages":"53","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"54","messages":"55","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"56","messages":"57","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"58","messages":"59","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"60","messages":"61","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"62","messages":"63","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"64","messages":"65","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"66","messages":"67","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"68","messages":"69","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"70","messages":"71","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"72","messages":"73","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"74","messages":"75","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"76","messages":"77","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"78","messages":"79","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"80","messages":"81","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/wleaverton/dev/jwplayer-fandom/src/i18n.ts",[],"/Users/wleaverton/dev/jwplayer-fandom/src/main.ts",[],"/Users/wleaverton/dev/jwplayer-fandom/src/players/DesktopArticleVideoPlayer/DesktopArticleVideoPlayer.tsx",[],"/Users/wleaverton/dev/jwplayer-fandom/src/players/DesktopArticleVideoPlayer/UnmuteButton.tsx",[],"/Users/wleaverton/dev/jwplayer-fandom/src/players/DesktopArticleVideoPlayer/UserFeedback/CloseButton.tsx",[],"/Users/wleaverton/dev/jwplayer-fandom/src/players/DesktopArticleVideoPlayer/UserFeedback/ThumbDownButton.tsx",[],"/Users/wleaverton/dev/jwplayer-fandom/src/players/DesktopArticleVideoPlayer/UserFeedback/ThumbUpButton.tsx",[],"/Users/wleaverton/dev/jwplayer-fandom/src/players/DesktopArticleVideoPlayer/UserFeedback/UserFeedback.tsx",[],"/Users/wleaverton/dev/jwplayer-fandom/src/players/DesktopArticleVideoPlayer/VideoDetails.tsx",[],"/Users/wleaverton/dev/jwplayer-fandom/src/players/VideoPlayer.tsx",[],"/Users/wleaverton/dev/jwplayer-fandom/src/players/shared/JwPlayerWrapper.tsx",[],"/Users/wleaverton/dev/jwplayer-fandom/src/players/shared/PlayerContext.ts",[],"/Users/wleaverton/dev/jwplayer-fandom/src/players/shared/PlayerWrapper.tsx",[],"/Users/wleaverton/dev/jwplayer-fandom/src/plugins/fandom-wirewax.plugin.ts",[],"/Users/wleaverton/dev/jwplayer-fandom/src/types.d.ts",[],"/Users/wleaverton/dev/jwplayer-fandom/src/utils/communication/communicationService.js",[],"/Users/wleaverton/dev/jwplayer-fandom/src/utils/communication/index.js",[],"/Users/wleaverton/dev/jwplayer-fandom/src/utils/communication/ofType.js",[],"/Users/wleaverton/dev/jwplayer-fandom/src/utils/useOnScreen.ts",[],"/Users/wleaverton/dev/jwplayer-fandom/src/players/shared/AdWrapper.tsx",[]]
